clean up some scala2013 testcases

+++ testcases/verification/datastructures/AssociativeList.scala
@@ -3,22 +3,22 @@ import leon.annotation._
 
 object AssociativeList { 
   sealed abstract class KeyValuePairAbs
-  case class KeyValuePair(key: Int, value: Int) extends KeyValuePairAbs
+  case class KeyValuePair(key: BigInt, value: BigInt) extends KeyValuePairAbs
 
   sealed abstract class List
   case class Cons(head: KeyValuePairAbs, tail: List) extends List
   case class Nil() extends List
 
   sealed abstract class OptionInt
-  case class Some(i: Int) extends OptionInt
+  case class Some(i: BigInt) extends OptionInt
   case class None() extends OptionInt
 
-  def domain(l: List): Set[Int] = l match {
-    case Nil() => Set.empty[Int]
+  def domain(l: List): Set[BigInt] = l match {
+    case Nil() => Set.empty[BigInt]
     case Cons(KeyValuePair(k,_), xs) => Set(k) ++ domain(xs)
   }
 
-  def find(l: List, e: Int): OptionInt = l match {
+  def find(l: List, e: BigInt): OptionInt = l match {
     case Nil() => None()
     case Cons(KeyValuePair(k, v), xs) => if (k == e) Some(v) else find(xs, e)
   }
@@ -39,11 +39,11 @@ object AssociativeList {
       case KeyValuePair(ek, ev) => if (ek == k) Cons(KeyValuePair(ek, ev), xs) else Cons(KeyValuePair(k, v), updateElem(xs, e))
     }
   }) ensuring(res => e match {
-    case KeyValuePair(k, v) => domain(res) == domain(l) ++ Set[Int](k)
+    case KeyValuePair(k, v) => domain(res) == domain(l) ++ Set[BigInt](k)
   })
 
   @induct
-  def readOverWrite(l: List, k1: Int, k2: Int, e: Int) : Boolean = {
+  def readOverWrite(l: List, k1: BigInt, k2: BigInt, e: BigInt) : Boolean = {
     find(updateElem(l, KeyValuePair(k2,e)), k1) == (if (k1 == k2) Some(e) else find(l, k1))
   } holds
 }
